dev-requirements.txt
liac-arff
bentoml-plugins-arize
sklearn
xgboost
pandas
numpy